requirements.txt
Flask==2.3.3
Werkzeug==3.0.3
joblib==1.4.2
pandas==2.2.2
scikit-learn==1.5.0
gunicorn==20.1.0
numpy==2.0.0
google-cloud-storage==2.17.0
python-dotenv==1.0.1
functions-framework==3.7.0